requirements.txt
crewai==0.1.0  
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.2
requests==2.28.1
python-dotenv==0.21.0  
